createUser.php & verify.php created
Added:
CreateUser.php which will insert new entrys into the data base for new
users basic information.
Verify.php used for login checks if username and password match up and
provideds a cookie allowing access to other parts of the website. Also
sends the user to the login page if they do not have a valid cookie.

Changed:
Tester.html added in another for to allow for testing of the
createuser.php script.