Create Signout.php & loadUser.js
Add:
Signout.php which removes the users login cookie and sends them to the
login screen.
loadUser.js which runs an Ajax call to get all the information about the
logged in user. Then adds the results into divs with specific id's.
popup.js when id called popin is clicked the div with id popup will
appear in the center of screen.