Major update
Added:
JS:
Validate Plugin imported onto the register page to validate form before
submission is allowed to occur.

registerValidation.js provides the rules that the form must follow to be
considered valid.

tooltipster.js is a precreated js program designed to highlight missing
parts in a form when it is submitted and not valid

PHP:
alreadyLoggedIn.php, if the user tries to go to the login screen and is
already logged in this php script will redirect the user to the profile
page.

validCookie.php checks if the user has a valid login cookie if it does
it sets a variable $verified to TRUE. Was created to take repeated code
out of other php files

Edited:
php:
delete.php has been edited to use validateCookie.php reducing the amount
of lines present in file

getInfo.php edited to use validateCookie.php reducing the amount of
lines present in file.

updateUser.php added commentation

JS:
loadUser.js edited to work with updated getInfo.php will try reload the
page if an error occurs during ajax request.