Grade Validator
Added:
PHP:
validateGrade.php which check if the addGrade form has been correctly
filled and that it is a valid qualification.

JS:
storeGrade.js which when the user hits a button passes the variables in
the form to validateGrade if its a valid grade it will store the values
in an array and display them at the bottom of the page for the user.
More functions to be implemented soon.