Improved storeGrade.js
storeGrades.js now check if the user has selected a course, level and
grade if not informs the user.  If they have checks that they are valid
and then inserts it into an array. Then when the submit button is
presses checks if the currently select info in the form is in the array.
If it is not it adds it to the array and inserts each qualification 1 at
a time into the database